Moved symmetric.h/symmetric.c inside of crypto subfolder. Some header reorganization and cleanup.
