+++ src/lib/crypto/symmetric.c
