+++ src/librepgp/stream-key.c
@@ -25,26 +25,26 @@
  */
 
 #include "config.h"
-#include "stream-def.h"
-#include "stream-key.h"
-#include "stream-armor.h"
-#include "stream-packet.h"
 #include <sys/stat.h>
 #include <stdlib.h>
 #include <stdio.h>
 #include <unistd.h>
 #include <string.h>
 #include <time.h>
+#include "stream-def.h"
+#include "stream-key.h"
+#include "stream-armor.h"
+#include "stream-packet.h"
 #include "defs.h"
 #include "types.h"
-#include "symmetric.h"
+#include "crypto/symmetric.h"
+#include "crypto/s2k.h"
 #include "fingerprint.h"
 #include "pgp-key.h"
 #include "list.h"
 #include "packet-parse.h"
 #include "utils.h"
 #include "crypto.h"
-#include "crypto/s2k.h"
 
 static void
 signature_list_destroy(list *sigs)
